+++ src/main/webapp/css/admin/adminLanTrip.css
@@ -1,9 +1,6 @@
 @charset "UTF-8";
 
-*{
-    margin: 0;
-    padding:0;
-}
+
 a{
     text-decoration: none;
 }
@@ -16,44 +13,9 @@ ul{
 }
 /* content 랩 */
 .wrap{
-    width : 1080px;
-    margin: auto;
+    width : 1200px;
 }
-/* 홈페이지 기능 네비 */ 
-.top_gnb_area{
-    width: 100%;
-    height: 50px;
-    background-color: #f0f0f1;
-    position:relative;
-}
-.top_gnb_area .list{
-    position: absolute;
-    top: 0px;
-    right: 0;
-    
-}
-.top_gnb_area .list li{
-    list-style: none;    
-    float : left;
-    padding: 13px 15px 0 10px;
-    font-weight: 900;
-    cursor: pointer;
-}
- 
-/* 관리자 페이지 상단 현페이지 정보 */
-.admin_top_wrap{
-    height:110px;
-    line-height: 110px;
-    background-color: #5080bd;
-    margin-bottom:15px;
-}
-.admin_top_wrap>span{
-    margin-left: 30px;
-    display:inline-block;
-    color: white;
-    font-size: 50px;
-    font-weight: bolder;
-}
+
 /* 관리자 wrap(네비+컨텐츠) */
 .admin_wrap{
     
@@ -93,26 +55,37 @@ ul{
  
 /* 관리자페이지 컨텐츠 영역 */
 .admin_content_wrap{
-    width: 80%;
     float:left;
+    width: 79%;
     height: 100%;
-    height: 700px;
-    
-}
-.admin_content_subject{            /* 관리자 컨텐츠 제목 영역 */
-    font-size: 20px;
-    font-weight: bolder;
-    padding-left: 15px;
-    background-color: #6AAFE6;
-    height: 40px;
-    line-height: 40px;
-    color: white;    
+    border-left: 1px solid gray;
+	padding: 5px;
+	margin-bottom: 10px; 
 }
 
-.admin_content_goods{            /* 항공 컨텐츠 상세 영역 */
+.admin_content_top {				/* 관리자 컨텐츠 제목 영역 */
+	background-color: #58CCFF;
+    padding: 10px;
+}
+
+.admin_content_subject{            /* 랜선여행 관리 제목*/
+    font-size: 20px;
+    font-weight: bolder;
+    color: white;
+    display: inline-block;
+    width: 60%;
+}
+
+.admin_content_goods_submit{            /* 등록, 수정, 삭제*/
+    display: inline-block;
+    padding: 5px;
+    background-color: #58CCFF;
+    border: none;
+}
+
+.admin_content_goods{            /* 컨텐츠 상세 영역 */
     height: 500px;
     padding-left: 15px;
-    background-color: skyblue;
     margin-bottom: 20px;
 }
 
@@ -121,76 +94,11 @@ ul{
 	background-color: #58CCFF;
 }
 
-.admin_content_goods_submit{            /* 항공 컨텐츠 submit */
-    height: 100px;
-    float:left;
-    background-color:white;
-    margin-left: 200px;
+.lan_list_table {
+	width: 100%;
+	margin: 10px;
+	padding: 5px;
 }
- 
-/* footer navai 영역 */
-.footer_nav{
-    width:100%;
-    height:50px;
-}
-.footer_nav_container{
-    width: 100%;
-    height: 100%;
-    background-color:#8EC0E4;
-}
-.footer_nav_container>ul{
-    font-weight : bold;
-    float:left;
-    list-style:none;
-    position:relative;
-    padding-top:10px;
-    line-height: 27px;
-    font-family: dotum;
-    margin-left: 350px;
-}
-.footer_nav_container>ul>li{
-    display:inline;
-    width: 45px;
-    height: 19px;
-    padding: 10px 9px 0 10px;
-}
-.footer_nav_container>ul>span{
-    margin: 0 4px;
-}
-/* footer 영역 */
-.footer{
-    width:100%;
-    height:130px;
-    background-color:#D4DFE6;
-    padding-bottom : 50px;
-}
-.footer_container{
-    width: 100%;
-    height: 100%;
-    margin: auto;
-}
-.footer_left>img {
-    width: 150%;
-    height: 130px;
-    margin-left: -20px;
-    margin-top: -12px;
-}
-.footer_left{
-    float :left;
-    width: 170px;
-    margin-left: 20px;
-    margin-top : 30px;
-    
-}
-.footer_right{
-    float :left;
-    width: 680px;
-    margin-left: 200px;
-    margin-top : 30px;
-}
- 
- 
- 
  
 /* float 속성 해제 */
 .clearfix{
